+++ CHANGELOG.md
@@ -6,12 +6,15 @@
 
 - Level select menu has been changed to allow for any number of levels which can be added without adding UI elements.
 - Added goal egg to end of existing levels, edited level to accomodate and increase linearity.
+- Enemies' player detection is more advanced, using a polygon set up in the tilemap rather than a using a line set up in code.
 
 ### Added
 
 - New level goal: reaching an egg sprite at the end of the level.
 - Added Player sprites correlating to flap charge.
 - Added cling sprite.
+- Added UI showing number of enemies defeated.
+- Added UI showing punch charge.
 
 ### Removed
 ### Fixed